Update dcservo.ino
motor was energized even when destination was reached.
for some reason the PID results had a residual PID value keeping the motor energized even when at destination.   The result was a hot motor idle at destination.  This way, the motor does not draw any power when error is zero but draws power when it is disturbed.